requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-cors-headers==4.3.1
django-model-utils==4.3.1
djangorestframework==3.14.0
google-auth==2.25.1
idna==3.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
